requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg[binary]>=3.1.18
alembic>=1.13.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
